requirements.txt
vtracer
Pillow